adding basic views for cart
the download cart allows a user to add datasets as they browse the website
a "checkout" page then allows the user to download all the datasets at once.
