+++ yeastphenome/apps/common/templates/base/navbar.html
@@ -20,6 +20,12 @@
             <li {% if 'conditions' == active %} class="active" {% endif %} ><a href="{% url 'conditions:index' %}">Experimental conditions</a></li>
             <li {% if 'about' == active %} class="active" {% endif %} ><a href="{% url 'common:about' %}">About</a></li>
             <li {% if 'contributors' == active %} class="active" {% endif %} ><a href="{% url 'contributors' %}">Contributors</a></li>
+            <li>
+               <a href="{% url 'common:view_cart' %}" class="nav-link waves-effect">
+               <i class="fa fa-shopping-cart" aria-hidden="true"></i>
+                    <span class="clearfix d-none d-sm-inline-block"></span>
+               </a>
+            </li>
         </ul>
 
         <div class="collapse navbar-collapse" id="nb-search">
@@ -29,6 +35,7 @@
                 </div>
                 <button type="submit" class="btn btn-default">Submit</button>
             </form>
+
         </div>
     </div>
 </nav>
