+++ yeastphenome/apps/common/templates/messages/message.html
@@ -1,0 +1,12 @@
+<div class="row">
+    <div class="col-md-12">
+      {% for message in messages %}
+        <div class="alert alert-{{ message.tags }} alert-dismissible" role="alert">
+          {{ message }}
+          <button type="button" class="close" data-dismiss="alert" aria-label="Close">
+            <span aria-hidden="true">&times;</span>
+          </button>
+        </div>
+        {% endfor %}
+    </div>
+</div>
