+++ yeastphenome/apps/common/utils.py
@@ -1,0 +1,263 @@
+from django.db.models import Q, F, Count, Sum, Case, When
+from django.db import models
+
+import numpy as np
+
+
+from yeastphenome.apps.papers.models import Paper
+from yeastphenome.apps.conditions.models import ConditionSet, ConditionType
+from yeastphenome.apps.phenotypes.models import Phenotype
+from yeastphenome.apps.datasets.models import Dataset
+
+
+# Stats helper function
+
+
+def get_latest_stats():
+    """Return number of papers, phenotypes, and datasets to display in the index
+       view. If no entries are found, display counts of zero.
+    """
+    papers_queryset = Paper.objects.all()
+    phenotypes_queryset = Phenotype.objects.all()
+    conditions_queryset = ConditionSet.objects.all()
+    datasets_queryset = Dataset.objects.all()
+
+    # Total number of papers to process
+    f = Q(latest_data_status__status__is_valid=True)
+    papers_nr = papers_queryset.filter(f).count()
+
+    # Latest modified paper
+    updated_on = papers_queryset.latest().modified_on if papers_queryset else None
+
+    # Number of hopeless papers
+    f = Q(latest_data_status__status__name__in=["request abandoned", "not available"])
+    papers_hopeless_nr = papers_queryset.filter(f).count()
+
+    # Number of labs
+    f = Q(latest_data_status__status__name__in=["not relevant"])
+    labs_nr = papers_queryset.exclude(f).values("last_author").distinct().count()
+
+    # Number of papers processed and loaded
+    f = Q(latest_data_status__status__name__exact="loaded") & Q(
+        latest_tested_status__status__name__in=[
+            "loaded",
+            "request abandoned",
+            "not available",
+        ]
+    )
+    papers_queryset = papers_queryset.filter(f)
+    papers_processed_nr = papers_queryset.count()
+
+    # Number of phenotypes
+    f = Q(dataset__paper__in=papers_queryset)
+    phenotypes_nr = phenotypes_queryset.filter(f).distinct().count()
+
+    # Number of conditions
+    f = Q(dataset__paper__in=papers_queryset)
+    conditions_nr = conditions_queryset.filter(f).distinct().count()
+
+    # Number of datasets (total)
+    f = Q(paper__in=papers_queryset)
+    datasets_nr = datasets_queryset.filter(f).distinct().count()
+
+    # --- Conditions ---
+    conditiontypes_qs = ConditionType.objects.all()
+    top_conditiontypes = (
+        conditiontypes_qs.annotate(
+            nr_papers=Count("condition__conditionset__dataset__paper", distinct=True)
+        )
+        .annotate(
+            nr_datasets=Sum(
+                Case(
+                    When(
+                        condition__conditionset__dataset__data_available__shortname__in=[
+                            "q",
+                            "qofh",
+                            "d",
+                        ],
+                        then=1,
+                    ),
+                    default=0,
+                    output_field=models.IntegerField(),
+                )
+            ),
+            nr_datasets_q=Sum(
+                Case(
+                    When(
+                        condition__conditionset__dataset__data_available__shortname="q",
+                        then=1,
+                    ),
+                    default=0,
+                    output_field=models.IntegerField(),
+                )
+            ),
+            nr_datasets_d=Sum(
+                Case(
+                    When(
+                        condition__conditionset__dataset__data_available__shortname="d",
+                        then=1,
+                    ),
+                    default=0,
+                    output_field=models.IntegerField(),
+                )
+            ),
+            nr_datasets_qofh=Sum(
+                Case(
+                    When(
+                        condition__conditionset__dataset__data_available__shortname="qofh",
+                        then=1,
+                    ),
+                    default=0,
+                    output_field=models.IntegerField(),
+                )
+            ),
+        )
+        .exclude(name__in=["standard", "time"])
+        .order_by("-nr_papers")
+    )
+
+    # top_conditiontypes_q = top_conditiontypes. \
+    #     annotate(nr_datasets_q=Count(condition__conditionset__dataset__paper__data_available__shortname='q'))
+
+    # --- Collections ----
+    f = Q(paper__in=papers_queryset)
+
+    c = Q(collection__shortname__in=["hap a"])
+    datasets_nr_hap_a = datasets_queryset.filter(f & c).distinct().count()
+    datasets_prc_hap_a = (
+        int(np.rint(100 * datasets_nr_hap_a / datasets_nr)) if datasets_nr != 0 else 0
+    )
+
+    c = Q(collection__shortname__in=["hap alpha"])
+    datasets_nr_hap_alpha = datasets_queryset.filter(f & c).distinct().count()
+    datasets_prc_hap_alpha = (
+        int(np.rint(100 * datasets_nr_hap_alpha / datasets_nr))
+        if datasets_nr != 0
+        else 0
+    )
+
+    c = Q(collection__shortname__in=["hom"])
+    datasets_nr_hom = datasets_queryset.filter(f & c).distinct().count()
+    datasets_prc_hom = (
+        int(np.rint(100 * datasets_nr_hom / datasets_nr)) if datasets_nr != 0 else 0
+    )
+
+    c = Q(collection__shortname__in=["het"])
+    datasets_nr_het = datasets_queryset.filter(f & c).distinct().count()
+    datasets_prc_het = (
+        int(np.rint(100 * datasets_nr_het / datasets_nr)) if datasets_nr != 0 else 0
+    )
+
+    c = Q(
+        collection__shortname__in=[
+            "hap ?",
+            "hap a/hap alpha/hom",
+            "hap a/hap alpha",
+            "hap a/hom",
+            "hom/het?",
+            "hom/het",
+            "hap a/het",
+            "hap ?/hom/het",
+        ]
+    )
+    datasets_nr_mix = datasets_queryset.filter(f & c).distinct().count()
+    datasets_prc_mix = (
+        int(np.rint(100 * datasets_nr_mix / datasets_nr)) if datasets_nr != 0 else 0
+    )
+
+    datasets_nr_collections_total = (
+        datasets_nr_hap_a
+        + datasets_nr_hap_alpha
+        + datasets_nr_hom
+        + datasets_nr_het
+        + datasets_nr_mix
+    )
+
+    # c = Q(collection__shortname__in=['hap a', 'hap alpha', 'hom', 'het', 'hap ?', 'hap a/hap alpha/hom',
+    #                                  'hap a/hap alpha', 'hap a/hom', 'hom/het?',
+    #                                  'hom/het', 'hap a/het', 'hap ?/hom/het'])
+    # missing = datasets_queryset.filter(f).exclude(c)
+
+    # --- Data types ---
+    f = Q(paper__in=papers_queryset) & Q(data_available__shortname="q")
+    datasets_nr_q = datasets_queryset.filter(f).distinct().count()
+    datasets_prc_q = (
+        int(np.rint(100 * datasets_nr_q / datasets_nr)) if datasets_nr != 0 else 0
+    )
+
+    f = Q(paper__in=papers_queryset) & Q(data_available__shortname="qofh")
+    datasets_nr_qofh = datasets_queryset.filter(f).distinct().count()
+    datasets_prc_qofh = (
+        int(np.rint(100 * datasets_nr_qofh / datasets_nr)) if datasets_nr != 0 else 0
+    )
+
+    f = Q(paper__in=papers_queryset) & Q(data_available__shortname="d")
+    datasets_nr_d = datasets_queryset.filter(f).distinct().count()
+    datasets_prc_d = (
+        int(np.rint(100 * datasets_nr_d / datasets_nr)) if datasets_nr != 0 else 0
+    )
+
+    datasets_nr_data_available_total = datasets_nr_q + datasets_nr_qofh + datasets_nr_d
+
+    # Data recovery for haploid/homozygous diploid
+    f = Q(paper__in=papers_queryset)
+
+    g1 = Q(
+        data_measured__rank__lt=F("data_published__rank")
+    )  # datasets in need of data recovery
+    g2 = Q(
+        data_available__rank__lt=F("data_published__rank")
+    )  # datasets with recovered data
+
+    h1 = Q(tested_list_published=False)  # datasets in need of tested list recovery
+    h2 = Q(tested_list_published=False) & Q(
+        tested_source_id__isnull=False
+    )  # datasets with recovered tested list
+
+    fgh = f & (g2 | h2)  # all datasets with something recovered
+
+    datasets_nr_need_data = datasets_queryset.filter(f & g1).distinct().count()
+    datasets_nr_need_tested = datasets_queryset.filter(f & h1).distinct().count()
+
+    datasets_nr_recovered_all = datasets_queryset.filter(fgh).distinct().count()
+    datasets_nr_recovered_data = datasets_queryset.filter(f & g2).distinct().count()
+    datasets_nr_recovered_tested = datasets_queryset.filter(f & h2).distinct().count()
+
+    # Heterozygous
+
+    context = {
+        "papers_nr": papers_nr,
+        "papers_hopeless_nr": papers_hopeless_nr,
+        "labs_nr": labs_nr,
+        "papers_processed_nr": papers_processed_nr,
+        "phenotypes_nr": phenotypes_nr,
+        "conditions_nr": conditions_nr,
+        "datasets_nr": datasets_nr,
+        "datasets_nr_hap_a": datasets_nr_hap_a,
+        "datasets_prc_hap_a": datasets_prc_hap_a,
+        "datasets_nr_hap_alpha": datasets_nr_hap_alpha,
+        "datasets_prc_hap_alpha": datasets_prc_hap_alpha,
+        "datasets_nr_hom": datasets_nr_hom,
+        "datasets_prc_hom": datasets_prc_hom,
+        "datasets_nr_het": datasets_nr_het,
+        "datasets_prc_het": datasets_prc_het,
+        "datasets_nr_mix": datasets_nr_mix,
+        "datasets_prc_mix": datasets_prc_mix,
+        "datasets_nr_collections_total": datasets_nr_collections_total,
+        "datasets_nr_q": datasets_nr_q,
+        "datasets_prc_q": datasets_prc_q,
+        "datasets_nr_qofh": datasets_nr_qofh,
+        "datasets_prc_qofh": datasets_prc_qofh,
+        "datasets_nr_d": datasets_nr_d,
+        "datasets_prc_d": datasets_prc_d,
+        "datasets_nr_data_available_total": datasets_nr_data_available_total,
+        "datasets_nr_need_data": datasets_nr_need_data,
+        "datasets_nr_need_tested": datasets_nr_need_tested,
+        "datasets_nr_recovered_all": datasets_nr_recovered_all,
+        "datasets_nr_recovered_data": datasets_nr_recovered_data,
+        "datasets_nr_recovered_tested": datasets_nr_recovered_tested,
+        "top_conditiontypes": top_conditiontypes[:10],
+        "updated_on": updated_on,
+    }
+
+    return context
