+++ yeastphenome/apps/datasets/templates/datasets/datasets_table.html
@@ -10,13 +10,13 @@
     </p>
 </div>
 
-{% if datasets %}
+{% include "messages/message.html" %}
 
-{% if template == 'paper' %}
+{% if datasets %}{% if template == 'paper' %}
 <a href="/{{ template }}s/{{ object.id }}/{{ DOWNLOAD_PREFIX }}_{{ object.pmid }}_datasets_list.txt">Download the list of datasets</a>
 {% endif %}
 
-<form action="/datasets/download/" method="get">
+<form action="{% url 'datasets:download' %}" method="get">
     <table id="papersTable" class="tablesorter table table-condensed table-hover" width="100%">
 
         <thead>
@@ -32,12 +32,11 @@
         </thead>
 
         <tbody>
-        {% for dataset in datasets %}
-        <tr>
+        {% for dataset in datasets %}<tr>
             <td><input id="{{ dataset.id }}" type="checkbox" name="{{ dataset.id }}" class="dataset" {# {% if not dataset.has_data_in_db %}disabled{% endif %} #}></td>
             <td class="col-md-2"><a href="{% url 'papers:detail' dataset.paper.id %}">{{ dataset.paper }}</a></td>
-            <td class="col-md-3">{{ dataset.phenotype.observable2.name }} {% if dataset.phenotype.reporter %} ({{ dataset.phenotype.reporter }}) {% endif %}</td>
-            <td class="col-md-2">{{ dataset.conditionset.display_name }}</td>
+            <td class="col-md-2">{{ dataset.phenotype.observable2.name }} {% if dataset.phenotype.reporter %} ({{ dataset.phenotype.reporter }}) {% endif %}</td>
+            <td class="col-md-1">{{ dataset.conditionset.display_name }}</td>
             <td class="col-md-2">{{ dataset.medium.display_name }}</td>
             <td class="col-md-1">{{ dataset.collection.shortname }}</td>
 
@@ -52,13 +51,11 @@
                 <span class="text-muted">{{ dataset.data_available|capfirst }}</span>
                 {% endif %}
             </td>
-
             <td>
-                <a href="/datasets/{{ dataset.id }}/"><span class="glyphicon glyphicon-plus" aria-hidden="true"></span></a>
+                <a href="/datasets/{{ dataset.id }}/"><button type="button" class="btn btn-primary btn-sm">Details</button></a>
+                {% if dataset.id not in request.session.cart %}<a href="{% url 'common:add_to_cart' dataset.id "datasets:growth" %}"><button type="button" class="btn btn-primary btn-sm" style="width:100px">Add to Cart</button></a>{% else %}<a href="{% url 'common:remove_from_cart' dataset.id "datasets:growth" %}"><button type="button" class="btn btn-danger btn-sm" style="width:100px">Remove</button></a>{% endif %}
             </td>
-        </tr>
-
-        {% endfor %}
+        </tr>{% endfor %}
         </tbody>
     </table>
